modification of email verification
The emailID is still generated from the email.
The list of emailIDs now associates ID to its email.

Therefore if the emailID is in the list, the onCreateAccount function also verifies
that the email corresponding to the ID is the email that generated the ID.

It frees the authentication system from the necessity to have @ensimag email adress,
but if the list of emailIDs only contains @ensimag email adresses, it is an equivalent
authntication system.