Allowed all grenoble-inp.org emails to create an account
Being in the electoral register is no longer required for creating an account, in order to ease the acount creating especially at the beginning of the school year.
Being in the electoral register is still required in order to vote!

Also updated from Node 8 to 10
Node 8 is being deprecated. But Node 10 requires Firebase blaze account